Adds a circular deque (to make controlgris2 real-time safer) (#3)
Add a "circular deque" a custom data structure

It was tailor made to replace the usage of std::deque in the audio thread of controlGRIS2. This is backed by a std::array so it does not allocate on push. 

Has not been tested in a multithreaded environment and is probably not thread safe.